+++ mission/chapter02/mission2/src/main.jsx
@@ -1,7 +1,6 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import App from './App.jsx'
-import './index.css'
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
